chore(main): small fixes, detailed below
Change things around to use const, include my ft_word sub-library, and
go to the next word before executing a builtin. This will probably be
changed around in the future, maybe if I use a linked list or something
like that I won't need to manage this main input string like this. First
it'll go through validating (to see if I need another line), then it'll
go through parsing (to expand everything), then it should be good to be
interpreted.