chore(environment): make function that takes namelen
get_env_i uses ft_strlen to get the length, which means it assumes the string
is null terminated, which usually means it's allocated as it's own thing,
which is usually not the case. So I've made a new function that takes the
length of the name to compare, so I can just use ft_strchr to find the =
or something (or when validating the expansions, I can look for the first
non-variable character).