refactor(validate_env): change return value
I've changed this function to return 1 on success, and 0 on failure. I
have also made a function that validates an environment variable name
with a length. This is so I can easily check if a variable name is valid
or not when I have a certain amount of characters to check for