refactor(args): expand command to linked list
Add functions to make a linked list of arguments from a command string.
It makes a linked list for each argument while it's expanding them, to
prevent recopying the string many times. once the argument has been
expanded into a list, it is converted into a full string and inserted
into the arguments list.

This commit also simplifies a bit of the tilde expanding, since it no
longer has to worry about chunking a section before the tilde.