chore(prompt): prompt selector jump table
My prompt printing function now uses a jump table to print the correct
prompt. It gets passed an index, validates it, and then runs the prompt
function the index refers to. I changed the comments too, and removed
the idea of a 4th prompt for now. So far it's going to be normal, single
quote, double quote (1, 2, and 3). I might make a generic continuation
prompt and be able to pass it the type of continuation, which I think
would be nice. I fixed prompts 2 and 3, they didn't have a comma after
the format string and were printing really strange stuff.