fix(dollar): expand without {}
I had set up the expansions to work with ${VAR}, which was pretty easy. This
commit adds expanding for variables that aren't surrounded by {}. It uses
a function that figures out the length of a valid env name starting at a
pointer, and with that length I try to find a matching environment
variable. If one wasn't found, it is skipped. If there's no variable after
the $, it does nothing (i.e. leaves the $)