fix(expansions): save length of arg
I'll need the length of the args when combining them with echo. I've just
realized that this is pretty stupid though, since I could just do strlen
when I need to instead of every single time. I'll fix that in a minute
to save the length when it's making the arg list, that way I don't have
to actually calculate it again