feat(exit): add exit command
Reorganize some of the main file, break out the main loop into a new function
to clean up some stuff, and give me space for the exit checking and appropriate
error handling. I had to set ret = 0 at the end of the shell_loop function,
because in the case that we're continuing a line, ret doesn't get reset at
the gnl call (because it doesn't make it that far in the while condition)
and so if for some reason it was set to -1 by process_command, it would
trigger the get_next_line error message.