chore(set_env): sort set_env functions
Three functions each for setting an environment variable to an int or a string.
One takes a name string, one takes an environment variable string with the length
of the name at the beginning, and the last takes an environment variable index.
This saves me from having to calculate the same thing multiple times in different
places.