refactor(get_env_value): merge multiple files
Rename to get_env_value, delete get_env and get_envn files.

Change the name scheme to get_env_value and get_envn_value. I have other
getting functions now, so this was necessary to prevent confusion.

Same as the set env functions: one that takes just the name, one that
takes a string and the length of the name at the beginning, and one that
takes a validated index. This is just to prevent getting the same
information multiple times.