feat(get_envi_int): get integer value from env
This doesn't check if the string is actually an integer, because it wouldn't
be able to return something sensible to say that it won't work (any possible
return value would be valid). So use ft_str_is_int from my libft on the
env value before doing this. (This is the reason that this function takes
an env_index, because before running it I've already found the index and
validated that the string is in fact an integer. I didn't want to have
this function need to find the index all over again, and I don't yet need
one that does all that, so for now there's only this one that takes an
index that has already been found.)