feat(shlvl): increase shlvl
This commit adds an env_inc and envn_inc function, which increment the integer
variable inside an environment variable. It's used to increment the shell level
variable. Also this commit adds a void shell_setup function, which does things
like incrementing shlvl, setting the shell variable, and printing the prompt
to start with. the env_inc functions need to be moved still.

squash fix(shell var): don't change

The SHELL variable, I found out, actually represents the login shell
of the user, so new terminals made use that shell. It does not represent
the currently running shell, so I should not change it.