refactor(inc_env): move to new file, clean up
env_inc changed to inc_env, and it's set up a bit differently from the
others. Instead of having the normal inc_envi, I went with a static
support function to make the incremented string instead, because that
made more sense. There was no way to get around data being re-calculated
if it was done the other way, so I'm happy that this way does't have that
problem. I also doubt I'll need to be able to increment an environment
variable based only on the index, and if I do, then it'll be really easy
to add that later.