fix(call_command): properly exit child process
When execve fails, it doesn't switch processes. Meaning, it leaves off
in the child right after the execve line. Before, I'd just return(1),
not really expecting it to happen, but I've had to change that to exit,
and I've also printed an error message if that happens. I've also added
an error message if fork fails (really I just combined it into one line,
because I used to (void) the print_error call and then return 1, but
since print_error returns 1 I just combined them. I do the same thing in
the call to exit()).