fix(permissions): properly check access to command
Set up a proper access call to check if we can run a command. We check
if it's readable and executable, because if it isn't both of those then
execve fails. Also add a nice comment describing what we're doing. Also
also, move the double array creation to inside the successful case,
which solves a lot of headaches with deleting it properly.