refactor(builtins.c): move functions elsewhere
check_builtins has been moved to it's own file, with a name change to
get_builtin_index, because that's what it does -- returns the index of
the builtin in the builtin strings (and the jump table). It is used to
verify that a builtin exists. It has also changed a little bit, instead
of taking a string with arguments for the command, it now takes just the
command, which means we can remove the wordequ call and replace it with
strnequ, also caching the length to avoid recalculating it each time.
The comment has been updated to reflect these changes.

run_builtin is now a static function inside process_command.c, so that
it can't be used outside of there. This is better, because this is the
only spot this jump table should be being used, so this stops other
people from potentially messing with it.

Comments have been added in get_builtin_index and run_builtin to ensure
that the tables are kept up to date with eachother. For now, this is
what I will do, but I imagine I'll find a better way to handle this
later.

I also changed the name of the builtin_nr variable to builtin_index in
process_command, this follows other conventions and makes more sense.