feat(error): properly print builtin name in error
Add a function that prints the builtin name as well as the error message. This
is necessary because usually error messages say the name of the function that's
erroring, and builtins also need the name of the shell. So in order for an error
like 'minishell: cd: /home/mint: Permission denied', I'd need to set up a format
string in my regular error printer (turning it into ft_printfd(2, ...), which
I do not want to do), or maybe use string literals for parts of it or something.
Instead of those bad ideas, I just made an error printer that takes 3 arguments.