feat(error codes)
This commit changes all error message printing to use error codes
instead. In places like ft_cd, where it had helper functions that
printed the error message, they now instead return the error code to be
taken care of in the main function. This is better, because this commit
also changes builtins to have the first arg in the list be the name of
the function, so that error messages can be printed in a more generic
way (i.e. we don't have to type the name of the command in the error
message anymore, so that way if anything changes, it changes
everywhere).

ft_setenv was changed to print a proper error message for having an
invalid environment variable, not just the name

ft_echo was changed to make the string differently, it now accepts no
arguments and just prints a newline. I had to add an if statement to see
if there was no length and if there were no links (i.e. nothing was
inputted), and then increase the size by 1 so the newline didn't
overwrite the null terminator. I couldn't increase links if it was zero,
because it uses that in the while loop. I also couldn't not check links,
because it's possible (in the future) to have arguments of null string
(just quotes with nothing inside them), and it would need to be spaced
properly with links, so this way I can be sure by checking both that
there are indeed no parameters.

A few functions in ft_unsetenv (and maybe elsewhere) were changed to
call the envn version, because I already had the length of the string
with args->content_size. this saves a recalculation.

This is a large commit, forgive me. Lots of things were changed that
depended on eachother, and I didn't want to separate everything.