fix(dollar expansions): check before chunking
When expanding env names with dollar signs, I used to be making
too many chunks (it'd make one at the beginning even if it was
empty, and it'd make one for the variable it grabbed, even if it
was empty). I have added some checks (and removed the comment
reminding me to do that).