fix(builtins): verify whole name instead of just part
Change strnequ to wordnequ, which actually makes sure the length of the
strings are the same. With strnequ, it only cared if the first characters
were the same (e.g. 'e' would match with 'echo', 'c' would match
'cd'). This way I can cache the length of the command, and then
for each builtin it gets compared against, wordnequ gets it's length
as well to make sure they're the same, and then it checks all
characters.