chore(gitignore): remove binary from gitignore
The main reason I did this, is because it made the minishell folder disappear
from netrw when I got it to hide things based on my gitignore file. That was
annoying, so I took out minishell from the .gitignore. Now it also kinda serves
as an indicator that the project is compiled, which is nice.