feat(prompt): add time and directory into prompt
The time isn't real yet, but it will be once I figure out how ctime
works. It's seg faulting for some reason when I've been trying it, so I
need to play with it still. I don't want to use a bunch of functions, so
hopefully what I'm trying to do is easy.