fix(commands): look in current folder if no PATH
While searching the path for an executable, this sets the path to "." if
there isn't one (so we search in the current folder). Also, if there are
any forward slashes in the command, instead of just duplicating that
string, I first check to see if there's an executable there (so that way
"No command found" gets printed instead of "Permission denied", when
later I assume that it exists and stat it to see if I have permissions).